customer-support/src/main/java/com/zlatkosh/customersupport/CustomerSupportController.java
package com.zlatkosh.customersupport;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseBody;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.server.ResponseStatusException;

@RequiredArgsConstructor
@RestController
@RequestMapping("/customer-support")
@Slf4j
public class CustomerSupportController {

    private final CustomerSupportService customerSupportService;

    @ApiResponses(value = {
            @ApiResponse(responseCode = "200", description = "Successfully completed the check",
                    content = @Content),
            @ApiResponse(responseCode = "500", description = "Internal Server Error",
                    content = @Content),
    })
    @Operation(description = "This controller is used to check if Customer support is available at the moment. ")
    @GetMapping("/check-status")
    @ResponseBody
    public CustomerSupportStatus checkStatus() {
        try {
            return customerSupportService.checkStatus();
        } catch (Exception e) {
            log.error("Internal Server Error: ", e);
            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Internal Server Error: %s".formatted(e.getMessage()));
        }
    }
}
